requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
requests==2.31.0
firebase-admin==6.4.0
google-generativeai==0.7.0
eventlet==0.35.2
gunicorn==21.2.0
